chore: bump toolhive and refactor automation openapi
